+++ goldens/api/@betterer/tsquery.d.ts
@@ -1,1 +1,3 @@
 export declare function tsquery(configFilePath: string, query: string): BettererFileTest;
+
+export declare function tsqueryΔ(query: string): BettererFileTest;