+++ packages/betterer/src/config/public.ts
@@ -15,6 +15,7 @@ export {
   BettererOptionsStartBase,
   BettererOptionsStartCI,
   BettererOptionsStartDefault,
+  BettererOptionsStartPrecommit,
   BettererOptionsStartStrict,
   BettererOptionsStartUpdate,
   BettererOptionsStart,