+++ packages/cli/src/start.ts
@@ -1,24 +1,17 @@
 import { BettererSummary, betterer, BettererOptionsStart } from '@betterer/betterer';
+import { ciΔ } from './ci';
 
-import { startOptions } from './options';
+import { cliOptions } from './options';
 import { BettererCLIArguments } from './types';
 
 /** @internal Definitely not stable! Please don't use! */
 export function startΔ(cwd: string, argv: BettererCLIArguments): Promise<BettererSummary> {
-  const {
-    cache,
-    cachePath,
-    config,
-    exclude,
-    filter,
-    include,
-    results,
-    reporter,
-    silent,
-    strict,
-    tsconfig,
-    update
-  } = startOptions(argv);
+  if (process.env.CI) {
+    return ciΔ(cwd, argv);
+  }
+
+  const { cache, cachePath, config, exclude, filter, include, results, reporter, silent, strict, tsconfig, update } =
+    cliOptions(argv);
 
   // Mark options as unknown...
   const options: unknown = {