+++ packages/tsquery/src/tsquery.ts
@@ -34,3 +34,34 @@ export function tsquery(configFilePath: string, query: string): BettererFileTest
     );
   });
 }
+
+/** @internal Definitely not stable! Please don't use! */
+export function tsqueryΔ(query: string): BettererFileTest {
+  if (!query) {
+    throw new BettererError(
+      "for `@betterer/tsquery` to work, you need to provide a query, e.g. `'CallExpression > PropertyAccessExpression'`. ❌"
+    );
+  }
+
+  const resolver = new BettererFileResolver();
+  return new BettererFileTest(resolver, async (filePaths, fileTestResult) => {
+    if (filePaths.length === 0) {
+      return;
+    }
+
+    await Promise.all(
+      filePaths.map(async (filePath) => {
+        const fileText = await fs.readFile(filePath, 'utf8');
+        const sourceFile = tsq.ast(fileText);
+        const matches = tsq.query(sourceFile, query, { visitAllChildren: true });
+        if (matches.length === 0) {
+          return;
+        }
+        const file = fileTestResult.addFile(filePath, fileText);
+        matches.forEach((match) => {
+          file.addIssue(match.getStart(), match.getEnd(), 'TSQuery match');
+        });
+      })
+    );
+  });
+}