+++ test/setup.ts
@@ -1,3 +1,4 @@
 import '@betterer/fixture';
+import { jest } from '@jest/globals';
 
 jest.setTimeout(200000);